requirements.txt
PyYAML
requests
feedparser
